# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - authentication_error
    - authorization_error
    - database_error
    - forbidden
    - internal_error
    - invalid_request
    - malformed_body
    - not_implemented
    - password_too_short
    - refresh_token_invalid
    - resource_already_exists
    - resource_not_found
    - token_invalid
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceNotFound
    - ErrCodeTokenInvalid
...
  database.GetUserByIDRow:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      user_role:
        type: string
    type: object
  database.GetUsersRow:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      updated_at:
        type: string
      user_role:
        type: string
    type: object
...
  handlers.CreateUserRequest:
    properties:
      email:
        example: example@example.com
        type: string
      password:
        description: passwords must be at least 11 characters long
        example: lkIB53@6O^Y
        type: string
    type: object
...
  handlers.UpdatePasswordRequest:
    properties:
      current_password:
        example: lkIB53@6O^Y
        type: string
      new_password:
        example: ue6U>&X3j570
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        example: example_error_code
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Signals ISN service API
paths:
...
  /admin/users/{id}:
    get:
      description: This API is protected (includes email addresses in the response)
        - currently only available on dev envs, pending implementation of admin roles.
      parameters:
      - description: user id
        example: 68fb5f5b-e3f5-4a96-8d35-cd2203a06f73
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.GetUserByIDRow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get registered user
      tags:
      - auth
...
  /api/users:
    get:
      description: This api displays email addresses and is currently only available
        on dev env pending implementation of role based access
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.GetUsersRow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get the registered users
      tags:
      - auth
  /auth/admins/{account_id}:
    delete:
      description: this endpoint can only be used by the site owner account
      parameters:
      - description: account id
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        in: path
        name: account_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Revoke account admin role
      tags:
      - auth
  /auth/admins/account/{account_id}:
    put:
      description: |-
        This endpoint grants the admin role to a site member

        An admin can:
        - Create an ISN
        - Define the signal_types used in the ISN
        - read/write to their own ISNs
        - Grant other accounts read or write access to their ISNs
        - Create service identities ('service accounts')

        this endpoint can only be used by the site owner account
      parameters:
      - description: account id
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        in: path
        name: account_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Grant admin role
      tags:
      - auth
...
  /auth/password/reset:
    put:
      description: |-
        Use this api to reset the user's password.  Requires a valid access token and the current password

        TODO - forgotten password facility
      parameters:
      - description: user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePasswordRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Reset password
      tags:
      - auth
  /auth/register:
    post:
      description: |-
        The first user to be created for this service will be created with an admin role.
        Subsequent accounts default to standard user roles.
      parameters:
      - description: user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create user
      tags:
      - auth
...
produces:
- application/json
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...